% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTRUE_vec}
\alias{isTRUE_vec}
\title{Internal: vectorised isTRUE that handles NA gracefully}
\usage{
isTRUE_vec(x)
}
\arguments{
\item{x}{A logical (or coercible) vector.}
}
\value{
Logical vector.
}
\description{
Returns a logical vector of the same length as \code{x}. Values that are
\code{NA} are returned as \code{FALSE}. Use this when filtering data.table
rows by a logical column that may contain \code{NA}.
}
\keyword{internal}
