% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{pdactrace-palettes}
\alias{pdactrace-palettes}
\alias{pdactrace_pal_class}
\alias{pdactrace_pal_group}
\alias{pdactrace_pal_dir}
\alias{pdactrace_pal_pattern}
\title{Palettes for pdactrace figures}
\format{
An object of class \code{character} of length 4.

An object of class \code{character} of length 4.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.
}
\usage{
pdactrace_pal_class

pdactrace_pal_group

pdactrace_pal_dir

pdactrace_pal_pattern
}
\description{
Named character vectors for ggplot2 \verb{scale_*_manual()}.
\itemize{
\item \code{pdactrace_pal_class}    - Class A / B / Other
\item \code{pdactrace_pal_group}    - HC / Pancreatitis / PDAC
\item \code{pdactrace_pal_dir}      - UP / DOWN / NS
\item \code{pdactrace_pal_pattern}  - Early × 4 pattern colors (atlas-surfaced subset of the v0.4.0 12-template catalog)
}
}
\examples{
pdactrace_pal_class
pdactrace_pal_pattern
}
\keyword{datasets}
