% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sb_average_precision}
\alias{sb_average_precision}
\title{Average precision}
\usage{
sb_average_precision(truth, score)
}
\arguments{
\item{truth}{Logical or binary vector.}

\item{score}{Numeric score; larger values predict positives.}
}
\description{
Average precision
}
