% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sb_binary_auc}
\alias{sb_binary_auc}
\title{Binary area under the ROC curve}
\usage{
sb_binary_auc(truth, score)
}
\arguments{
\item{truth}{Logical or binary vector.}

\item{score}{Numeric score; larger values predict positives.}
}
\description{
Binary area under the ROC curve
}
